webkit/data/layout_tests/platform/chromium-win/LayoutTests/fast/leaks/002-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,<input type='file' id='input'></input> from frame with URL 002.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 17: Uncaught TypeError: Cannot call method 'getElementById' of undefined
This test exercises RenderBlock::bidiReorderCharacters in a specific way to test for BidiRun leaks as reported in rdar://problem/4987649. Its layout has no particular meaning.
